// src/angular/src/app/interfaces/cours.ts
export interface Cours{
    numero: string;  
    id: string;   
    duree: string;  
    cumul: string;  
    ddebut: string;
    dfin : string;  
    nbsem: string;  
    jour : string;  
    hdebut: string;
    hfin : string;  
    typecours : string;  
    codemat: string;
    matiere: string;
    enseignant: string;
    codeenseignant: string;
    codediplome: string;
    libdiplome: string;
    codesalle: string;  
    salle: string;  
    pond : string;  
    prop : string;  
    memo : string;  
    date_modif: string;
}